// src/movies/infraestructure/repository/gateway/dto.ts
export interface SwapiMovie {
  title: string;
  episode_id: number;
  opening_crawl: string;
  director: string;
  producer: string;
  release_date: string;
  characters: string[];
  planets: string[];
  starships: string[];
  vehicles: string[];
  species: string[];
  created: Date;
  edited: Date;
  url: string;
}

export interface ResponseSwapi {
  count: number;
  next: any;
  previous: any;
  results: SwapiMovie[];
}